Added separate tmp for each run of the pipeline...
Since any folder can only hold 100 000 files, this is necessary on UPPMAX. At leaste if more than one instance of the pipeline is to run at any on time. Also added functionality which ensures that this tmp directory is removed as long as the script dies gracefully. Scripts being kill with a -9 will still require a manual clean up however. But this should be rare as scancel on uppmax does not send this signal per default.